introduces the command line interfaces of the workflow and job
